test: temporary fix for - test_delete_exemptions_invalid_id

+++ services/accounts/settings.go
@@ -2,6 +2,7 @@ package accounts
 
 import (
 	"context"
+	"fmt"
 	"time"
 
 	"github.com/status-im/status-go/multiaccounts/accounts"
@@ -186,6 +187,9 @@ func (api *SettingsAPI) NotificationsSetExemptions(id string, muteAllMessages bo
 }
 
 func (api *SettingsAPI) DeleteExemptions(id string) error {
+	if id == "" {
+		return fmt.Errorf("invalid exemption id: cannot be empty")
+	}
 	return api.db.DeleteExemptions(id)
 }
